+++ package.json
@@ -35,6 +35,7 @@
     "@testing-library/user-event": "^14.4.3",
     "eslint-plugin-jest": "^27.2.1",
     "jest": "^29.4.3",
+    "jest-canvas-mock": "^2.5.0",
     "jest-environment-jsdom": "^29.4.3"
   },
   "description": "Next.js Template (incl. Styled Components and Jest)"
